add Epikcord.ext.commands and implement sections
