+++ EpikCord/__init__.py
@@ -10,6 +10,7 @@ from typing import *
 from urllib.parse import quote
 import io
 import os
+from .ext.commands import Section
 
 CT = TypeVar('CT', bound='Colour')
 T = TypeVar('T')
@@ -1702,25 +1703,7 @@ class HTTPClient(ClientSession):
 
         return res
 
-class Section:
-    def __init__(self):
-        self.commands = {}
-        self.events = {}
 
-    def event(self, event_name: str):
-        def register_event(func):
-            self.events[event_name] = func
-        return register_event
-
-    def slash_command(self, *, name: str, description: Optional[str], options: List[Union[Subcommand, SubCommandGroup, StringOption, IntegerOption, BooleanOption, UserOption, ChannelOption, RoleOption, MentionableOption, NumberOption]]):
-        def register_slash_command(func):
-            self.commands[name] = {
-                "callback": func,
-                "name": name,
-                "description": description,
-                "options": options
-            }
-        return register_slash_command
 
 
 class MissingClientSetting(Exception):
