+++ EpikCord/ext/commands/__init__.py
@@ -1,0 +1,1 @@
+from .sections import *
