+++ setup.cfg
@@ -24,6 +24,7 @@ include_package_data = True
 packages = 
     EpikCord
     EpikCord.managers
+    EpikCord.ext.commands
 install_requires = 
     aiohttp; python_version == ">=3.8"
 
@@ -32,3 +33,4 @@ install_requires =
 
 [options.extras_require]
 voice = PyNaCl
+speed = aiodns,brotli,orjson
